Split file handling of ArrayDefinitionSource to new PHPFileDefinitionSource
