+++ src/DI/Definition/Source/ArrayDefinitionSource.php
@@ -10,13 +10,12 @@
 namespace DI\Definition\Source;
 
 use DI\Definition\Definition;
-use DI\Definition\Exception\DefinitionException;
 use DI\Definition\MergeableDefinition;
 use DI\Definition\ValueDefinition;
 use DI\DefinitionHelper\DefinitionHelper;
 
 /**
- * Reads DI definitions from a PHP array, or a file returning a PHP array.
+ * Reads DI definitions from a PHP array.
  *
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
@@ -28,44 +27,16 @@ class ArrayDefinitionSource implements ChainableDefinitionSource
     private $chainedSource;
 
     /**
-     * @var bool
-     */
-    private $initialized;
-
-    /**
-     * File containing definitions, or null if the definitions are given as a PHP array.
-     * @var string|null
-     */
-    private $file;
-
-    /**
      * DI definitions in a PHP array
      * @var array
      */
     private $definitions = array();
 
     /**
-     * @param string|null $file File in which the definitions are returned as an array.
-     */
-    public function __construct($file = null)
-    {
-        if (! $file) {
-            $this->initialized = true;
-            return;
-        }
-
-        // If we are given a file containing an array, we lazy-load it to improve performance
-        $this->initialized = false;
-        $this->file = $file;
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function getDefinition($name, MergeableDefinition $parentDefinition = null)
     {
-        $this->initialize();
-
         if (! array_key_exists($name, $this->definitions)) {
             // Not found, we use the chain or return null
             if ($this->chainedSource) {
@@ -119,31 +90,6 @@ class ArrayDefinitionSource implements ChainableDefinitionSource
     public function addDefinition(Definition $definition)
     {
         $this->definitions[$definition->getName()] = $definition;
-    }
-
-    /**
-     * Lazy-loading of the definitions.
-     * @throws DefinitionException
-     */
-    private function initialize()
-    {
-        if ($this->initialized === true) {
-            return;
-        }
-
-        if (! is_readable($this->file)) {
-            throw new DefinitionException("File {$this->file} doesn't exist or is not readable");
-        }
-
-        $definitions = require $this->file;
-
-        if (! is_array($definitions)) {
-            throw new DefinitionException("File {$this->file} should return an array of definitions");
-        }
-
-        $this->addDefinitions($definitions);
-
-        $this->initialized = true;
     }
 
     /**
